Replace deprecated numpy aliases with builtin types

+++ trackeval/baselines/stp.py
@@ -49,7 +49,7 @@ def track_sequence(seq_file):
 
         # Initialize container for holding previously tracked objects.
         prev = {'boxes': np.empty((0, 4)),
-                'ids': np.array([], np.int),
+                'ids': np.array([], int),
                 'timesteps': np.array([])}
 
         # Run tracker for each timestep.
@@ -80,7 +80,7 @@ def track_sequence(seq_file):
             match_cols = match_cols[actually_matched_mask]
 
             # Assign the prev track ID to the current dets if they were matched.
-            ids = np.nan * np.ones((len(boxes),), np.int)
+            ids = np.nan * np.ones((len(boxes),), int)
             ids[match_cols] = prev['ids'][match_rows]
 
             # Create new track IDs for dets that were not matched to previous tracks.
